fabrique-js/requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
idna==2.7
Markdown==2.1.0
Pillow==5.3.0
pkg-resources==0.0.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
sorl-thumbnail==12.5.0
stripe==2.10.0
urllib3==1.23
